requirements.txt
# Note: this requirements.txt file is used to specify what dependencies are
# needed to launch the testing
setuptools
wheel
jmespath
PyYAML>=5.1
lxml
pycdlib
netaddr
pywinrm
pypsrp
pyvmomi
passlib
pytesseract
Pillow
ansible-core
